+++ src/main/java/coupon/repository/CouponRepository.java
@@ -1,4 +1,4 @@
-package coupon.service;
+package coupon.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
